pyproject.toml
[project]
name = "agentic-rag-gdpr-noyb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles==23.2.1",
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.11",
    "aiosignal==1.3.2",
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "asgiref==3.8.1",
    "attrs==25.1.0",
    "backoff==2.2.1",
    "bcrypt==4.2.1",
    "beautifulsoup4==4.12.3",
    "blinker==1.9.0",
    "build==1.2.2.post1",
    "cachetools==5.5.1",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "chroma-hnswlib==0.7.6",
    "chromadb==0.6.3",
    "click==8.1.8",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "dataclasses-json==0.6.7",
    "deprecated==1.2.18",
    "distro==1.9.0",
    "duckduckgo-search==7.3.0",
    "durationpy==0.9",
    "fastapi==0.115.8",
    "ffmpy==0.5.0",
    "filelock==3.17.0",
    "flatbuffers==25.1.24",
    "frozenlist==1.5.0",
    "fsspec==2024.12.0",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "google-auth==2.38.0",
    "googleapis-common-protos==1.66.0",
    "gradio==5.14.0",
    "gradio-client==1.7.0",
    "greenlet==3.1.1",
    "grpcio==1.70.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.27.2",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.28.1",
    "humanfriendly==10.0",
    "idna==3.10",
    "importlib-metadata==8.5.0",
    "importlib-resources==6.5.2",
    "jinja2==3.1.5",
    "jiter==0.8.2",
    "joblib==1.4.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "kubernetes==32.0.0",
    "langchain==0.3.17",
    "langchain-chroma==0.2.1",
    "langchain-community==0.3.16",
    "langchain-core==0.3.33",
    "langchain-huggingface==0.1.2",
    "langchain-text-splitters==0.3.5",
    "langsmith==0.3.3",
    "litellm==1.59.10",
    "lxml==5.3.0",
    "markdown-it-py==3.0.0",
    "markdownify==0.14.1",
    "markupsafe==2.1.5",
    "marshmallow==3.26.0",
    "mdurl==0.1.2",
    "mmh3==5.1.0",
    "monotonic==1.6",
    "mpmath==1.3.0",
    "multidict==6.1.0",
    "mypy-extensions==1.0.0",
    "narwhals==1.24.1",
    "networkx==3.4.2",
    "numpy==1.26.4",
    "oauthlib==3.2.2",
    "onnxruntime==1.20.1",
    "openai==1.60.2",
    "opentelemetry-api==1.29.0",
    "opentelemetry-exporter-otlp-proto-common==1.29.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.29.0",
    "opentelemetry-instrumentation==0.50b0",
    "opentelemetry-instrumentation-asgi==0.50b0",
    "opentelemetry-instrumentation-fastapi==0.50b0",
    "opentelemetry-proto==1.29.0",
    "opentelemetry-sdk==1.29.0",
    "opentelemetry-semantic-conventions==0.50b0",
    "opentelemetry-util-http==0.50b0",
    "orjson==3.10.15",
    "overrides==7.7.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "pillow==11.1.0",
    "posthog==3.11.0",
    "primp==0.11.0",
    "propcache==0.2.1",
    "protobuf==5.29.3",
    "pyarrow==19.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.1",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pydantic-settings==2.7.1",
    "pydeck==0.9.1",
    "pydub==0.25.1",
    "pygments==2.19.1",
    "pypdf==5.2.0",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "pyreadline3==3.5.4",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
    "pytz==2025.1",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==13.9.4",
    "rpds-py==0.22.3",
    "rsa==4.9",
    "ruff==0.9.4",
    "safehttpx==0.1.6",
    "safetensors==0.5.2",
    "scikit-learn==1.6.1",
    "scipy==1.15.1",
    "semantic-version==2.10.0",
    "sentence-transformers==3.4.1",
    "setuptools==75.8.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.2",
    "smolagents==1.6.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "sqlalchemy==2.0.37",
    "starlette==0.45.3",
    "streamlit==1.41.1",
    "sympy==1.13.1",
    "tenacity==9.0.0",
    "threadpoolctl==3.5.0",
    "tiktoken==0.8.0",
    "tokenizers==0.21.0",
    "toml==0.10.2",
    "tomlkit==0.13.2",
    "torch==2.6.0",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "transformers==4.48.2",
    "typer==0.15.1",
    "typing-extensions==4.12.2",
    "typing-inspect==0.9.0",
    "tzdata==2025.1",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
    "watchdog==6.0.0",
    "watchfiles==1.0.4",
    "websocket-client==1.8.0",
    "websockets==14.2",
    "wrapt==1.17.2",
    "yarl==1.18.3",
    "zipp==3.21.0",
    "zstandard==0.23.0",
]
